Add web search tool example and fix README typo
- Created a new example in `examples/web-search/main.go` demonstrating
  how to use the `web_search` tool type with Chat Completion.
- Fixed a typo in `examples/README.md` where the path to examples was
  incorrectly listed as `./example/` instead of `./examples/`.
- Removed the embedding example as requested.
